requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
brotlipy=0.7.0=py38h1a28f6b_1002
ca-certificates=2022.4.26=hca03da5_0
certifi=2021.10.8=py38hca03da5_2
cffi=1.15.0=py38h22df2f2_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
conda=4.12.0=py38hca03da5_0
conda-content-trust=0.1.1=pyhd3eb1b0_0
conda-package-handling=1.8.1=py38h1a28f6b_0
cryptography=36.0.0=py38h834c97f_0
cycler=0.11.0=pypi_0
fonttools=4.34.4=pypi_0
idna=3.3=pyhd3eb1b0_0
kiwisolver=1.4.4=pypi_0
libcxx=12.0.0=hf6beb65_1
libffi=3.4.2=hc377ac9_2
matplotlib=3.5.2=pypi_0
ncurses=6.3=h1a28f6b_2
openssl=1.1.1n=h1a28f6b_0
packaging=21.3=pypi_0
pandas=1.4.3=pypi_0
pillow=9.2.0=pypi_0
pip=21.2.4=py38hca03da5_0
pycosat=0.6.3=py38h1a28f6b_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py38hca03da5_0
python=3.8.13=hbdb9e5c_0
python-dateutil=2.8.2=pypi_0
python.app=3=py38h1a28f6b_0
pytz=2022.1=pypi_0
readline=8.1.2=h1a28f6b_1
requests=2.27.1=pyhd3eb1b0_0
ruamel_yaml=0.15.100=py38h1a28f6b_0
setuptools=61.2.0=py38hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.3=h1058600_0
tk=8.6.11=hb8d0fd4_1
tqdm=4.64.0=py38hca03da5_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py38hca03da5_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1a28f6b_1
yaml=0.2.5=h1a28f6b_0
zlib=1.2.12=h5a0b063_2
